backend/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["kylin <260987762@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
django = "^5.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
